requirements.txt
Django==5.2.12
djangorestframework==3.17.0
